added comments, swapped functions order
